Made changes to ExtraFunctions class: 	Added method to calculate the norm of a vector of variable length using Java varargs
Made changes to SensorCalibrationFragment:
	Fixed the passing of stride_length to GraphActivity via Intent

Renamed class GyroIntegrate to GyroscopeDeltaOrientation:
	Renamed method getIntegratedValues() to calcDeltaOrientation()

Made changes to MagneticFieldOrientation:
	Rotating DCM representing heading so that heading is negative when moving East of Norrth
	Rotation DCM representing heading by -pi/2 so that heading of 0 corresponds with North

+++ app/src/main/java/nisargpatel/deadreckoning/dialog/SensorCalibrationDialogFragment.java
@@ -112,7 +112,7 @@ public class SensorCalibrationDialogFragment extends DialogFragment {
     private Intent addExtras(Intent myIntent) {
 
         myIntent.putExtra("user_name", getArguments().getString("user_name", "unknown"));
-        myIntent.putExtra("stride_length", getArguments().getDouble("stride_length", 2.5));
+        myIntent.putExtra("stride_length", getArguments().getFloat("stride_length", 2.5f));
         myIntent.putExtra("preferred_step_counter", getArguments().getString("preferred_step_counter", "default"));
         myIntent.putExtra("step_detector", getArguments().getBoolean("step_detector", false));
